refactor/move Logo component to ui directory
